Add files via upload
Using a Multi-layer Perceptron (MLP) architecture with 1290085 parameters, a Deep Learning Model is developed to predict the Grade Point Average (GPA) of students based on a variety of characteristics, such as the topic, instructor, assessments (quizzes, assignments, examinations), and student ID. The approach seeks to shed light on the connections between these variables and academic achievement, which might help teachers identify students who would benefit from extra help. It also allows students to focus on the particular subjects in which they are struggling.